requirements.txt
python-dotenv
groq
requests
Flask_cors
Flask